+++ src/calendar/index.js
@@ -304,12 +304,13 @@ class Calendar extends Component {
     }
 
     return (
-      <View 
+      <View
         style={[this.style.container, this.props.style]}
         accessibilityElementsHidden={this.props.accessibilityElementsHidden} // iOS
         importantForAccessibility={this.props.importantForAccessibility} // Android
       >
         <CalendarHeader
+          testID={this.props.testID}
           ref={c => this.header = c}
           style={this.props.headerStyle}
           theme={this.props.theme}